+++ packages/projection-lifecycle/src/index.ts
@@ -2,4 +2,6 @@ export * from "./lease.js";
 export * from "./materialize.js";
 export * from "./wal.js";
 export * from "./append.js";
+export * from "./close.js";
+export * from "./recovery.js";
 export * from "./lifecycle.js";